# backend/docs/swagger.yml
openapi: 3.0.3
info:
    title: Noteo API
    description: Api for working with notes
    version: 1.0.0
servers:
    - url: http://localhost:2000
...
paths:
    /auth/login:
        post:
            tags:
                - "Authorization"
            summary: Giving access to a personal account
            requestBody:
                required: true
                $ref: "#/components/requestBodies/login"
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    id:
                                        type: string
                                        example: glsdfbpsodw45hb
                                    token:
                                        type: string
                                        example: token string
                                    email:
                                        type: string
                                        example: example@ex.ex
                                    createdAt:
                                        type: string
                                        example: 12-12-2023T12:12:12.000Z
                                    updatedAt:
                                        type: string
                                        example: 12-12-2023T12:12:12.000Z | null
                    description: User is logged in
                "403":
                    description: data entered incorrectly
                    $ref: "#/components/responses/jsonMessageFromServer"
                "404":
                    description: user not found
                    $ref: "#/components/responses/jsonMessageFromServer"
    /auth/registration:
        post:
            tags:
                - "Authorization"
            summary: Create account in the system
            requestBody:
                required: true
                $ref: "#/components/requestBodies/registration"
            responses:
                "200":
                    description: User has create the account
                    $ref: "#/components/responses/jsonMessageFromServer"
                "403":
                    description: data entered incorrectly
                    $ref: "#/components/responses/jsonMessageFromServer"
                "404":
                    description: user not found
                    $ref: "#/components/responses/jsonMessageFromServer"
...
components:
    requestBodies:
        login:
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            email:
                                type: string
                                example: example@example.ex
                            password:
                                type: string
                                example: password
        registration:
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            email:
                                type: string
                                example: example@example.ex
                            password:
                                type: string
                                example: password
    responses:
...
        jsonMessageFromServer:
            description: Message schema
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
                                example: message